Fix: id Claim을 Integer -> Long으로 변환하도록 수정

+++ src/main/kotlin/team/goodpeople/security/jwt/JWTFilter.kt
@@ -45,7 +45,7 @@ class JWTFilter(
              * */
             val username = jwtUtil.getClaim(accessToken, "username")
             val role = jwtUtil.getClaim(accessToken, "role")
-            val userId = jwtUtil.getClaim(accessToken, "id", Long::class.java)
+            val userId = jwtUtil.getClaim(accessToken, "id", Integer::class.java).toLong()
 
             val user = User.createEntityForJWT(
                 id = userId,
